refactor(optchain): 补丁直接写入官方源码，删除整文件复制
将可选链补丁从 3rd/lua-patch/optchain/ 的整文件复制改为直接写入
3rd/lua{54,55} 官方源码，全部由 #if defined(BEE_OPTCHAIN) 门控：

- lparser.c / lvm.c / lopcodes.{h,c} / lopnames.h：每文件仅增几行门控
  补丁（lparser.c ~110 行、lvm.c 15 行、指令表各 3-7 行）
- OP_SETTOP 枚举/模式表/反汇编名均以 BEE_OPTCHAIN 门控，默认构建
  预处理后与官方源码字节级一致，零影响
- compile 不再切换 onelua.c 来源，仅保留 -optchain 时定义宏

分支相对 master 的 diff 从 ~10100 行降至 ~500 行。
默认 + optchain（lua54/lua55）全部测试通过。
