+++ 3rd/lua54/lopcodes.h
@@ -307,10 +307,17 @@ OP_VARARG,/*	A C	R[A], R[A+1], ..., R[A+C-2] = vararg		*/
 OP_VARARGPREP,/*A	(adjust vararg parameters)			*/
 
 OP_EXTRAARG/*	Ax	extra (larger) argument for previous opcode	*/
+#if defined(BEE_OPTCHAIN)
+,OP_SETTOP/*	A B	R[A], ..., R[A+B] := nil; top := A+B+1	*/
+#endif
 } OpCode;
 
 
+#if defined(BEE_OPTCHAIN)
+#define NUM_OPCODES	((int)(OP_SETTOP) + 1)
+#else
 #define NUM_OPCODES	((int)(OP_EXTRAARG) + 1)
+#endif
