+++ 3rd/lua55/lopcodes.c
@@ -106,6 +106,9 @@ LUAI_DDEF const lu_byte luaP_opmodes[NUM_OPCODES] = {
  ,opmode(0, 0, 0, 0, 0, iABx)		/* OP_ERRNNIL */
  ,opmode(0, 0, 1, 0, 1, iABC)		/* OP_VARARGPREP */
  ,opmode(0, 0, 0, 0, 0, iAx)		/* OP_EXTRAARG */
+#if defined(BEE_OPTCHAIN)
+ ,opmode(0, 0, 0, 0, 1, iABC)		/* OP_SETTOP */
+#endif
 };
