+++ 3rd/lua55/lopcodes.h
@@ -345,10 +345,17 @@ OP_ERRNNIL,/*	A Bx	raise error if R[A] ~= nil (K[Bx - 1] is global name)*/
 OP_VARARGPREP,/* 	(adjust varargs)				*/
 
 OP_EXTRAARG/*	Ax	extra (larger) argument for previous opcode	*/
+#if defined(BEE_OPTCHAIN)
+,OP_SETTOP/*	A B	R[A], ..., R[A+B] := nil; top := A+B+1	*/
+#endif
 } OpCode;
 
 
+#if defined(BEE_OPTCHAIN)
+#define NUM_OPCODES	((int)(OP_SETTOP) + 1)
+#else
 #define NUM_OPCODES	((int)(OP_EXTRAARG) + 1)
+#endif
